Change OrbitalHistory to regular array
Here's the thought: giant orbits require huge amounts of data if holding one state for each second. But most of that data is very very close together in physical space, making it relatively redundant. 

But much like transforming coordinate vectors to orbital elements simplifies the maths of the problem, transforming our indexing basis from time to equal-Mean-Anomaly-divisions mean that all orbits have a constant data size and we can use interpolation between points of the orbit if necessary.

This, in turn, means our data can be a much simpler statically-sized array. We can use STL algorithms to search for data in the array. We can define the comparison operator to mean compared in time.

+++ Davidian/Engine/OrbitalHistory.cpp
@@ -2,11 +2,12 @@
 
 namespace engine {
 
-OrbitalHistory::OrbitalHistory(const orbital::Orbit& orbit) {
+OrbitalHistory::OrbitalHistory(const orbital::Orbit& orbit)
+{
 
 }
 
-OrbitState OrbitalHistory::stateAtTime(size_t seconds) {
+OrbitState OrbitalHistory::stateAtTime(double seconds) {
   return OrbitState();
 }
